requirements.txt
python-telegram-bot==22.4
aiohttp==3.12.15
python-dotenv==1.1.1